Updates to improve pylint score
